+++ R/ggdonut.R
@@ -4,7 +4,8 @@
 #' @param group_key Column used to summarize the data. Default: NULL.
 #' @param count_type Data frame type, chosen from "count" and "full". "count" means summarized data and "full" means full data. Default: count.
 #' @param fill_color Colors used. Default: NULL (conduct automatic selection).
-#' @param label_info Label information type, chosen from count, ratio and all (count and ratio). Default: count.
+#' @param label_info Label information type, combine from group, count, ratio.
+#' For example, use "count" , "ratio" will show count and ratio, count is main label info, ratio is in brackets. Default: count.
 #' @param label_split Pattern used to split the label, support regular expression. Default: space.
 #' @param label_len The length of label text. Used when \code{label_split} is NULL. Default: 40.
 #' @param label_color Color of the label. Default: black.
@@ -43,51 +44,51 @@
 #' # circle label and out of pie
 #' ggdonut(
 #'   data = diamonds, group_key = "cut", count_type = "full",
-#'   label_info = "all", label_type = "circle",
+#'   label_info = c("count", "ratio"), label_type = "circle",
 #'   label_size = 4, label_pos = "out"
 #' )
 #' # circle label and in pie plot, with no split
 #' ggdonut(
 #'   data = diamonds, group_key = "cut", count_type = "full",
-#'   label_info = "all", label_type = "circle", label_split = NULL,
+#'   label_info = c("count", "ratio"), label_type = "circle", label_split = NULL,
 #'   label_size = 4, label_pos = "in"
 #' )
 #' # horizon label and in pie plot, with no split
 #' ggdonut(
 #'   data = diamonds, group_key = "cut", count_type = "full",
-#'   label_info = "all", label_type = "horizon", label_split = NULL,
+#'   label_info = c("count", "ratio"), label_type = "horizon", label_split = NULL,
 #'   label_size = 4, label_pos = "in"
 #' )
 #' # horizon label and in pie plot
 #' ggdonut(
 #'   data = diamonds, group_key = "cut", count_type = "full",
-#'   label_info = "all", label_type = "horizon",
+#'   label_info = c("count", "ratio"), label_type = "horizon",
 #'   label_size = 4, label_pos = "in"
 #' )
 #' # horizon label and out of pie plot, with no split
 #' ggdonut(
 #'   data = diamonds, group_key = "cut", count_type = "full",
-#'   label_info = "all", label_type = "horizon", label_split = NULL,
+#'   label_info = c("count", "ratio"), label_type = "horizon", label_split = NULL,
 #'   label_size = 4, label_pos = "out"
 #' )
 #' # horizon label and out of pie plot
 #' ggdonut(
 #'   data = diamonds, group_key = "cut", count_type = "full",
-#'   label_info = "all", label_type = "horizon",
+#'   label_info = c("count", "ratio"), label_type = "horizon",
 #'   label_size = 4, label_pos = "out"
 #' )
 #' # with label threshold
 #' ggdonut(
 #'   data = diamonds, group_key = "cut", count_type = "full",
-#'   label_info = "all", label_type = "horizon", label_split = NULL,
+#'   label_info = c("count", "ratio"), label_type = "horizon", label_split = NULL,
 #'   label_size = 4, label_pos = "in", label_threshold = 10
 #' )
 #' ggdonut(
 #'   data = diamonds, group_key = "cut", count_type = "full",
-#'   label_info = "all", label_type = "horizon",
+#'   label_info = c("count", "ratio"), label_type = "horizon",
 #'   label_size = 4, label_pos = "in", label_threshold = 10
 #' )
-ggdonut <- function(data, group_key = NULL, count_type = c("count", "full"), fill_color = NULL, label_info = c("count", "ratio", "all"),
+ggdonut <- function(data, group_key = NULL, count_type = c("count", "full"), fill_color = NULL, label_info = "count",
                     label_split = "[[:space:]]+", label_len = 40, label_color = "black",
                     label_type = c("circle", "horizon", "none"), label_pos = c("in", "out"), label_gap = 0.05,
                     label_threshold = NULL, label_size = 4, border_color = "black", border_size = 1,
@@ -95,7 +96,6 @@ ggdonut <- function(data, group_key = NULL, count_type = c("count", "full"), fil
                     nudge_x = 1, nudge_y = 1) {
   # check parameters
   count_type <- match.arg(arg = count_type)
-  label_info <- match.arg(arg = label_info)
   label_type <- match.arg(arg = label_type)
   label_pos <- match.arg(arg = label_pos)