+++ R/ggrosepie.R
@@ -4,7 +4,7 @@
 #' @param group_key Column used to summarize the data, one or two are acceptable. Default: NULL.
 #' @param count_type Data frame type, chosen from "count" and "full". "count" means summarized data and "full" means full data. Default: count.
 #' @param fill_color Colors used. When length of \code{group_key} is two, color the subgroup, otherwise the main group. Default: NULL (conduct automatic selection).
-#' @param label_info Label information type, chosen from count, ratio and all (count and ratio). Default: count.
+#' @param label_info Label information type of outer pie plot, chosen from count, ratio and all (count and ratio). Default: count.
 #' @param label_split Pattern used to split the label, support regular expression. Default: NULL.
 #' @param label_len The length of label text. Used when \code{label_split} is NULL. Default: 40.
 #' @param label_color Color of the label. When length of \code{group_key} is two, this should be set to one color. Default: black.
@@ -84,7 +84,7 @@ ggrosepie <- function(data, group_key = NULL, count_type = c("count", "full"), f
   if (length(group_key) == 1) {
     plot.data <- PrepareData(
       data = data, group_key = group_key, count_type = count_type, fill_color = fill_color,
-      label_info = label_info, label_split = NULL, label_len = NULL, label_color = label_color
+      label_info = "count", label_split = NULL, label_len = NULL, label_color = label_color
     )
     data <- plot.data[["data"]]
     fill_color <- plot.data[["fill_color"]]